requirements.txt
# Sanctum Crucible - Dependencies

# Core ML
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0

# Fine-tuning
peft>=0.10.0
trl>=0.8.0

# Model serving
openai>=1.12.0  # Client for vLLM's OpenAI-compatible API
# vllm installed separately in setup script (CUDA version dependent)

# Data & config
pyyaml>=6.0
requests>=2.31.0

# Garak (LLM vulnerability scanner)
# garak installed separately in setup script

# Analysis
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
